+++ xyui/avalonia/src/XYUI.Avalonia/Controls/XYUI1/XYUI1-06-Link/XYLink.cs
@@ -4,6 +4,11 @@ namespace XYUI.Avalonia.Controls;
 
 public sealed class XYLink : Button
 {
-    public XYLink() => Classes.Add("xyui-link");
+    public XYLink()
+    {
+        Classes.Add("xyui-1-component");
+        Classes.Add("xyui-link");
+        Classes.Add("xyui-focusable");
+    }
     public string CanonicalId => "XYUI-1-06";
 }